refactor: moves `promptConfig` and required function into `config.js`
